Added ability to set refresh progress bar in all activities through parent Activity. Added ability to show connection error message in all activities.
