+++ moVirt/src/main/res/layout/activity_host_detail.xml
@@ -1,15 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
-    <ProgressBar
-        android:id="@+id/progress"
-        style="@style/commonProgressStyle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content">
-    </ProgressBar>
+    <include layout="@layout/progress_bar" />
 
     <android.support.v4.view.ViewPager
         android:id="@+id/viewPager"
@@ -21,7 +16,7 @@
             android:id="@+id/pagerTabStrip"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_gravity="top"/>
+            android:layout_gravity="top" />
 
     </android.support.v4.view.ViewPager>
 </RelativeLayout>
