+++ moVirt/src/main/res/layout/activity_vm_detail.xml
@@ -1,20 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-android:orientation="vertical"
-android:layout_width="match_parent"
-android:layout_height="match_parent">
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
     <RelativeLayout
-        android:orientation="vertical"
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:orientation="vertical">
 
-        <ProgressBar
-            android:id="@+id/progress"
-            style="@style/commonProgressStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content">
-        </ProgressBar>
+        <include layout="@layout/progress_bar" />
 
         <android.support.v4.view.ViewPager
             android:id="@+id/viewPager"
@@ -26,7 +21,7 @@ android:layout_height="match_parent">
                 android:id="@+id/pagerTabStrip"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_gravity="top"/>
+                android:layout_gravity="top" />
 
         </android.support.v4.view.ViewPager>
     </RelativeLayout>
