+++ moVirt/src/main/res/layout/progress_bar.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<merge xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <ProgressBar
+        android:id="@+id/progress"
+        style="@style/commonProgressStyle"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content" />
+</merge>
